Update publish-new-version.yaml (#5691)
Add more details about the how the version should be formatted. 

When trying to run this workflow, I'm always very unsure whether the
version should be prepended with a `v` or not. This should be clearly
stated in the description. If it doesn't matter, that should also be
stated explicitly.

Updated the texts after talking to @chriswk and making a release.