chore: add `reason` column to change request schedule table (#5802)
This PR adds a new `reason` column to the change request schedules table
and populates it with the data that is in the `failure_reason` column.

This is the expand phase of the expand/contract pattern. The code in
enterprise will be updated to try and use the new column name, but fall
back to the old one if no value is present.

The old column can be removed later.