React examples doc with images and relative links (#5921)
<!-- Thanks for creating a PR! To make it easier for reviewers and
everyone else to understand what your changes relate to, please add some
relevant content to the headings below. Feel free to ignore or delete
sections that you don't think are relevant. Thank you! ❤️ -->

## About the changes
<!-- Describe the changes introduced. What are they and why are they
being introduced? Feel free to also add screenshots or steps to view the
changes if they're visual. -->

This is a large PR with all of the React examples building on top of the
original tutorial.

Based off of this Google Doc:
https://docs.google.com/document/d/1m3ujgyrkJTCGLsEsBr1_eq4HQKqsf-lfvGFDJobns44/edit

I also included updates to the React tutorial links so they are relative
paths, not full URL paths.

I propose we make fast-follow updates to this piece of documentation for
changes we need to make in order to release and iterate on minor
updates.

### Important files
<!-- PRs can contain a lot of changes, but not all changes are equally
important. Where should a reviewer start looking to get an overview of
the changes? Are any files particularly important? -->

website/docs/feature-flag-tutorials/react/examples.md

---------

Co-authored-by: Thomas Heartman <thomas@getunleash.io>