docs: fix navbar styling issues: GitHub icon, button background, hidden text (#6078)
This PR fixes the three issues in the nav bar:
1. The GitHub icon was rendered in black in light mode. It should be
white.
2. The toggle theme button's background was white in light mode. It
should be a darker green.
3. The navbar would intersect with the logo due to not compacting
correctly.

The first two issues were caused by using CSS nesting, which is
apparently not supported in Safari yet (and might only be supported if
you have enabled a flag in other browsers).

The last issue was fixed by changing the syntax of a media query. The
new syntax is unsupported in Safari.