fix: maintenance mode should assume disable if db call fails.
Usually maintenance mode is disabled. If the call throws, which we
see a lot of when a unleash instance is in terminating state, we
should return a defualt value.

By having it throw inside of the memoizee function, the response is
not cached, and it will trigger new calls until it return a cachable
result.