docs: hide the quickstart and doc nav bar links when you're on the docs pages
This change hides the quickstart and doc nav bar links when you're on
the docs pages.

This gets around the issue of both of those links being
highlighted when you're on the docs pages.o

This should work on  91.66% of browsers. For older browsers, the links
will still be visible.